+++ test/membrane/core/element/lifecycle_controller_test.exs
@@ -42,7 +42,7 @@ defmodule Membrane.Core.Element.LifecycleControllerTest do
       })
 
     state =
-      struct(State,
+      struct!(State,
         module: DummyElement,
         name: :test_name,
         type: :filter,
@@ -50,7 +50,9 @@ defmodule Membrane.Core.Element.LifecycleControllerTest do
         parent_pid: self(),
         synchronization: %{clock: nil, parent_clock: nil},
         handling_action?: false,
+        supplying_demand?: false,
         pads_to_snapshot: MapSet.new(),
+        delayed_demands: MapSet.new(),
         pads_data: %{
           input:
             struct(Membrane.Element.PadData,
@@ -63,7 +65,10 @@ defmodule Membrane.Core.Element.LifecycleControllerTest do
               input_queue: input_queue,
               demand: 0
             )
-        }
+        },
+        satisfied_auto_output_pads: MapSet.new(),
+        awaiting_auto_input_pads: MapSet.new(),
+        auto_input_pads: []
       )
 
     assert_received Message.new(:atomic_demand_increased, :some_pad)