+++ test/membrane/core/element/stream_format_controller_test.exs
@@ -35,15 +35,17 @@ defmodule Membrane.Core.Element.StreamFormatControllerTest do
       })
 
     state =
-      struct(State,
+      struct!(State,
         module: Filter,
         name: :test_name,
-        parent: self(),
         type: :filter,
         playback: :playing,
         synchronization: %{clock: nil, parent_clock: nil},
         handling_action?: false,
+        supplying_demand?: false,
         pads_to_snapshot: MapSet.new(),
+        delayed_demands: MapSet.new(),
+        handle_demand_loop_counter: 0,
         pads_data: %{
           input:
             struct(Membrane.Element.PadData,
@@ -54,7 +56,9 @@ defmodule Membrane.Core.Element.StreamFormatControllerTest do
               input_queue: input_queue,
               demand: 0
             )
-        }
+        },
+        satisfied_auto_output_pads: MapSet.new(),
+        awaiting_auto_input_pads: MapSet.new()
       )
 
     assert_received Message.new(:atomic_demand_increased, :some_pad)