+++ test/membrane/integration/auto_demands_test.exs
@@ -4,9 +4,12 @@ defmodule Membrane.Integration.AutoDemandsTest do
   import Membrane.ChildrenSpec
   import Membrane.Testing.Assertions
 
+  alias Membrane.Buffer
   alias Membrane.Testing.{Pipeline, Sink, Source}
 
-  defmodule AutoDemandFilter do
+  require Membrane.Pad, as: Pad
+
+  defmodule ExponentialAutoFilter do
     use Membrane.Filter
 
     def_input_pad :input, accepted_format: _any
@@ -33,6 +36,32 @@ defmodule Membrane.Integration.AutoDemandsTest do
         {[buffer: {:output, buffer}], %{state | counter: 1}}
       end
     end
+  end
+
+  defmodule NotifyingAutoFilter do
+    use Membrane.Filter
+
+    def_input_pad :input, accepted_format: _any, availability: :on_request
+    def_output_pad :output, accepted_format: _any
+
+    @impl true
+    def handle_playing(_ctx, state), do: {[notify_parent: :playing], state}
+
+    @impl true
+    def handle_parent_notification(actions, _ctx, state), do: {actions, state}
+
+    @impl true
+    def handle_buffer(pad, buffer, _ctx, state) do
+      actions = [
+        notify_parent: {:handling_buffer, pad, buffer},
+        buffer: {:output, buffer}
+      ]
+
+      {actions, state}
+    end
+
+    @impl true
+    def handle_end_of_stream(_pad, _ctx, state), do: {[], state}
   end
 
   defmodule AutoDemandTee do
@@ -64,7 +93,7 @@ defmodule Membrane.Integration.AutoDemandsTest do
           :down -> {mult_payloads, payloads}
         end
 
-      filter = %AutoDemandFilter{factor: factor, direction: direction}
+      filter = %ExponentialAutoFilter{factor: factor, direction: direction}
 
       pipeline =
         Pipeline.start_link_supervised!(
@@ -81,6 +110,8 @@ defmodule Membrane.Integration.AutoDemandsTest do
 
       assert_end_of_stream(pipeline, :sink)
       refute_sink_buffer(pipeline, :sink, _buffer, 0)
+
+      Pipeline.terminate(pipeline)
     end
   end)
 
@@ -108,6 +139,8 @@ defmodule Membrane.Integration.AutoDemandsTest do
     end)
 
     refute_sink_buffer(pipeline, :left_sink, %{payload: 25_000})
+
+    Pipeline.terminate(pipeline)
   end
 
   test "handle removed branch" do
@@ -127,6 +160,8 @@ defmodule Membrane.Integration.AutoDemandsTest do
       assert_sink_buffer(pipeline, :left_sink, buffer)
       assert buffer.payload == payload
     end)
+
+    Pipeline.terminate(pipeline)
   end
 
   defmodule NotifyingSink do
@@ -174,6 +209,8 @@ defmodule Membrane.Integration.AutoDemandsTest do
           {:buffer_arrived, %Membrane.Buffer{payload: ^payload}}
         )
       end
+
+      Pipeline.terminate(pipeline)
     end
   end)
 
@@ -202,7 +239,7 @@ defmodule Membrane.Integration.AutoDemandsTest do
       Pipeline.start_link_supervised!(
         spec:
           child(:source, PushSource)
-          |> child(:filter, AutoDemandFilter)
+          |> child(:filter, ExponentialAutoFilter)
           |> child(:sink, Sink)
       )
 
@@ -230,7 +267,7 @@ defmodule Membrane.Integration.AutoDemandsTest do
       Pipeline.start_supervised!(
         spec:
           child(:source, PushSource)
-          |> child(:filter, AutoDemandFilter)
+          |> child(:filter, ExponentialAutoFilter)
           |> child(:sink, %Sink{autodemand: false})
       )
 
@@ -244,6 +281,163 @@ defmodule Membrane.Integration.AutoDemandsTest do
     assert_receive(
       {:DOWN, _ref, :process, ^pipeline, {:membrane_child_crash, :sink, _sink_reason}}
     )
+  end
+
+  defp source_definiton(name) do
+    # Testing.Source fed with such actions generator will produce buffers with incremenal
+    # sequence of numbers as payloads
+    actions_generator =
+      fn counter, _size ->
+        Process.sleep(1)
+
+        buffer = %Buffer{
+          metadata: %{creator: name},
+          payload: counter
+        }
+
+        actions = [buffer: {:output, buffer}, redemand: :output]
+        {actions, counter + 1}
+      end
+
+    %Source{output: {1, actions_generator}}
+  end
+
+  defp setup_pipeline_with_notifying_auto_filter(_context) do
+    pipeline =
+      Pipeline.start_link_supervised!(
+        spec: [
+          child({:source, 0}, source_definiton({:source, 0}))
+          |> via_in(Pad.ref(:input, 0))
+          |> child(:filter, NotifyingAutoFilter)
+          |> child(:sink, %Sink{autodemand: false}),
+          child({:source, 1}, source_definiton({:source, 1}))
+          |> via_in(Pad.ref(:input, 1))
+          |> get_child(:filter)
+        ]
+      )
+
+    [pipeline: pipeline]
+  end
+
+  describe "auto flow queue" do
+    setup :setup_pipeline_with_notifying_auto_filter
+
+    defp receive_processed_buffers(pipeline, limit, acc \\ [])
+
+    defp receive_processed_buffers(_pipeline, limit, acc) when limit <= 0 do
+      Enum.reverse(acc)
+    end
+
+    defp receive_processed_buffers(pipeline, limit, acc) do
+      receive do
+        {Pipeline, ^pipeline,
+         {:handle_child_notification, {{:handling_buffer, _pad, buffer}, :filter}}} ->
+          receive_processed_buffers(pipeline, limit - 1, [buffer | acc])
+      after
+        500 -> Enum.reverse(acc)
+      end
+    end
+
+    test "when there is no demand on the output pad", %{pipeline: pipeline} do
+      manual_flow_queue_size = 40
+
+      assert_pipeline_notified(pipeline, :filter, :playing)
+
+      buffers = receive_processed_buffers(pipeline, 100)
+      assert length(buffers) == manual_flow_queue_size
+
+      demand = 10_000
+      Pipeline.message_child(pipeline, :sink, {:make_demand, demand})
+
+      buffers = receive_processed_buffers(pipeline, 2 * demand)
+      buffers_number = length(buffers)
+
+      # check if filter processed proper number of buffers
+      assert demand <= buffers_number
+      assert buffers_number <= demand + manual_flow_queue_size
+
+      # check if filter processed buffers from both sources
+      buffers_by_creator = Enum.group_by(buffers, & &1.metadata.creator)
+      assert Enum.count(buffers_by_creator) == 2
+
+      # check if filter balanced procesesd buffers by their origin - numbers of
+      # buffers coming from each source should be similar
+      counter_0 = Map.fetch!(buffers_by_creator, {:source, 0}) |> length()
+      counter_1 = Map.fetch!(buffers_by_creator, {:source, 1}) |> length()
+      sources_ratio = counter_0 / counter_1
+
+      assert 0.8 < sources_ratio and sources_ratio < 1.2
+
+      Pipeline.terminate(pipeline)
+    end
+
+    test "when an element returns :pause_auto_demand and :resume_auto_demand action", %{
+      pipeline: pipeline
+    } do
+      manual_flow_queue_size = 40
+      auto_flow_demand_size = 400
+
+      assert_pipeline_notified(pipeline, :filter, :playing)
+
+      Pipeline.message_child(pipeline, :filter, pause_auto_demand: Pad.ref(:input, 0))
+
+      # time for :filter to pause demand on Pad.ref(:input, 0)
+      Process.sleep(500)
+
+      buffers = receive_processed_buffers(pipeline, 100)
+      assert length(buffers) == manual_flow_queue_size
+
+      demand = 10_000
+      Pipeline.message_child(pipeline, :sink, {:make_demand, demand})
+
+      # fliter paused auto demand on Pad.ref(:input, 0), so it should receive
+      # at most auto_flow_demand_size buffers from there and rest of the buffers
+      # from Pad.ref(:input, 1)
+      buffers = receive_processed_buffers(pipeline, 2 * demand)
+      buffers_number = length(buffers)
+
+      assert demand <= buffers_number
+      assert buffers_number <= demand + manual_flow_queue_size
+
+      buffers_by_creator = Enum.group_by(buffers, & &1.metadata.creator)
+      counter_0 = Map.get(buffers_by_creator, {:source, 0}, []) |> length()
+      counter_1 = Map.fetch!(buffers_by_creator, {:source, 1}) |> length()
+
+      # at most auto_flow_demand_size buffers came from {:source, 0}
+      assert auto_flow_demand_size - manual_flow_queue_size <= counter_0
+      assert counter_0 <= auto_flow_demand_size
+
+      # rest of them came from {:source, 1}
+      assert demand - auto_flow_demand_size <= counter_1
+
+      Pipeline.message_child(pipeline, :filter, resume_auto_demand: Pad.ref(:input, 0))
+
+      # time for :filter to resume demand on Pad.ref(:input, 0)
+      Process.sleep(500)
+
+      Pipeline.message_child(pipeline, :sink, {:make_demand, demand})
+
+      buffers = receive_processed_buffers(pipeline, 2 * demand)
+      buffers_number = length(buffers)
+
+      # check if filter processed proper number of buffers
+      assert demand <= buffers_number
+      assert buffers_number <= demand + manual_flow_queue_size
+
+      # check if filter processed buffers from both sources
+      buffers_by_creator = Enum.group_by(buffers, & &1.metadata.creator)
+      assert Enum.count(buffers_by_creator) == 2
+
+      # check if filter balanced procesesd buffers by their origin - numbers of
+      # buffers coming from each source should be similar
+      counter_0 = Map.fetch!(buffers_by_creator, {:source, 0}) |> length()
+      counter_1 = Map.fetch!(buffers_by_creator, {:source, 1}) |> length()
+      sources_ratio = counter_0 / counter_1
+
+      assert 0.8 < sources_ratio and sources_ratio < 1.2
+
+      Pipeline.terminate(pipeline)
+    end
   end
 
   defp reduce_link(link, enum, fun) do